Updated Maps and added a few changes
I added some of my own maps, "What has become of me?" and "Resistance is Futile". I also added "Future Imperfect" to the list. Added the ability to open the prison doors in "Future Imperfect" with the hopes of making this an objective. In "Darkness in Sight", removed a ton of road to drive, freeing up space for more game logic. Lastly, added some camera logic for when Cortez is spotted in "Into Darkness"