Adding Character/Story Logic
Adding Logic to "Into Darkness", "Danger in Sight", "What Has Become of Me?", and "Resistance is Futile" with some logic. If you kill Time Assassin Cortez, it is an automatic failure to story. Added time crystal to objective in "Resistance is Futile". "What has Become of Me?" is intended to be a walk-through map, to end the story is to have Crow and Dr Cortez see the player. More story objectives/plot points are needed for a more cohesive story.